refactor: move authentication logic into dedicated controller
- Extracted authentication logic from routes to a separate authController
- Improved code organization and separation of concerns by implementing MVC architecture
- Enhanced maintainability and scalability of the authentication module
