+++ backend/controllers/authController.js
@@ -1,0 +1,60 @@
+/* eslint-disable no-unused-vars */
+/* eslint-disable no-undef */
+const jwt = require('jsonwebtoken');
+const bcrypt = require('bcrypt');
+const User = require('../models/User');
+const { v4: uuidv4 } = require('uuid');
+
+exports.register = async (req, res) => {
+  const { name, email, password } = req.body;
+
+  try {
+    // Vérifier si l'utilisateur existe déjà
+    const existingUser = await User.findOne({ $or: [{ email }, { name }] });
+    if (existingUser) {
+      return res.status(400).json({
+        message:
+          existingUser.email === email
+            ? 'Email is already taken'
+            : 'Name is already taken',
+      });
+    }
+
+    // Créer un nouvel utilisateur sans hacher le mot de passe ici
+    const user = new User({ name, email, password });
+    await user.save();
+
+    // Générer un token JWT
+    const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, {
+      expiresIn: '1h',
+    });
+    res.status(201).json({ token });
+  } catch (error) {
+    res.status(500).json({ message: 'Error creating user' });
+  }
+};
+
+exports.login = async (req, res) => {
+  const { email, password } = req.body;
+  try {
+    // Vérifier si l'utilisateur existe
+    const user = await User.findOne({ email });
+    if (!user) {
+      return res.status(404).json({ message: 'User not found' });
+    }
+
+    // Vérifier le mot de passe
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) {
+      return res.status(400).json({ message: 'Invalid credentials' });
+    }
+
+    // Générer un token JWT
+    const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, {
+      expiresIn: '1h',
+    });
+    res.status(200).json({ token });
+  } catch (error) {
+    res.status(500).json({ message: 'Error logging in' });
+  }
+};
